🔥 Remove custom serviceworker
